Policies

+++ app/Http/Controllers/SpecialWorkingHoursFactorController.php
@@ -5,18 +5,15 @@ namespace App\Http\Controllers;
 use App\Models\Organization;
 use App\Models\SpecialWorkingHoursFactor;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
-use Inertia\Inertia;
 
 class SpecialWorkingHoursFactorController extends Controller
 {
-    public function show(SpecialWorkingHoursFactor $specialWorkingHoursFactor)
-    {
-        return Inertia::render('SWHF/SWHFShow', ['specialWorkingHoursFactor' => $specialWorkingHoursFactor]);
-    }
-
     public function store(Request $request)
     {
+        Gate::authorize('create', SpecialWorkingHoursFactor::class);
+
         $validated = $request->validate([
             'type' => ['required', Rule::in(array_keys(SpecialWorkingHoursFactor::$TYPES))],
             'extra_charge' => 'required|numeric',
@@ -25,7 +22,6 @@ class SpecialWorkingHoursFactorController extends Controller
             'type.in' => 'Muss in ' . join(', ', SpecialWorkingHoursFactor::$TYPES) . ' vorhanden sein'
         ]);
 
-
         SpecialWorkingHoursFactor::updateOrCreate([
             'id' => $validated['id']
         ], [
@@ -33,13 +29,13 @@ class SpecialWorkingHoursFactorController extends Controller
             'organization_id' => Organization::getCurrent()->id
         ]);
 
-
         return back()->with('success', "Besonderer Arbeitszeitzuschlag erfolgreich gespeichert.");
     }
 
-    
-    public function destroy(Request $request, SpecialWorkingHoursFactor $specialWorkingHoursFactor)
+    public function destroy(SpecialWorkingHoursFactor $specialWorkingHoursFactor)
     {
+        Gate::authorize('delete', $specialWorkingHoursFactor);
+
         $specialWorkingHoursFactor->delete();
 
         return back()->with('success', 'Arbeitszuschlag erfolgreich gelöscht.');
